part_03-026_split_contains_av/src/Exercise026/Program.cs
namespace Exercise026
{
    using System;
    public class Program
    {
        public static void Main(string[] args)
        {
            
            while (true)
            {
                string str = Console.ReadLine();

                if (str == "")
                    break;

                string[] pieces = str.Split(" ");

                foreach (string piece in pieces)
                {
                    if (piece.Contains("av"))
                    {
                        Console.WriteLine(piece);
                    }

        }
    }
}

    }
}